batt_smbus: Adds check for if the battery detected is a 3DR solo battery
Also, if the battery is a 3DR Solo battery, it checks for the button
press on the battery which sets the is_powering_off flag.

The check for whether the battery is a 3DR Solo battery is based on
whether the Manufacturer Name, Device Name and Device Chemistry all
match the that of the standard 3DR Solo battery.

(cherry picked from commit 4b6d0ef0a6e0dc0a37212df728a2b74ce0637e87)

Change-Id: I9b5bd5ac076d363da187c83f5e5161f84acb709b